# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Simple POS API
  version: 1.0.0
  description: |-
    Contract-first API for Simple POS (ADR-0008). Precedence: BUSINESS_RULES.md (BR v1.2) > this file > the UI/UX PDF > code.

    **Hosts**: API `https://api.cashier.hossam-ameen.online`, PWA `https://cashier.hossam-ameen.online` (separate origins,
    CORS limited to the PWA origin, ADR-0003).

    **Conventions** (ADR-0005, 0008, 0010, 0013)
    - Paths have no trailing slash. JSON is snake_case. Ids are integers. Timestamps are RFC 3339 UTC.
    - **Money is always an integer number of minor units (piastres) in a field ending in `_minor`** (BR-GEN-01). The server
      computes every total, change, expected cash and difference (BR-GEN-03).
    - Success body: `{success: true, message, data}`; lists: `data = {items, pagination, …}`; 204 has no body.
    - Error body: `{success: false, code, message, details}`; `code` is a BR §9 code with its BR §9 HTTP status and Arabic
      `message` (BR-GEN-06).
    - Auth: `Authorization: Bearer <access_token>` (15 min). The refresh token is an httpOnly cookie on the API host only,
      used by `/api/auth/refresh` and `/api/auth/logout` with `X-Requested-With: simple-pos` (ADR-0003, ADR-0011).
    - `x-roles` lists the roles allowed to call each operation (`PUBLIC`, `ANY_AUTHENTICATED`, `ADMIN`, `CASHIER`); every
      other authenticated role gets `FORBIDDEN_ROLE` (BR-ROLE-05). `x-br` lists the rules the operation enforces and
...

...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /api/orders:
    get:
      operationId: listOrders
      tags:
      - orders
      summary: Order history (screen 15) and shift order lists (screens 02, 04)
      description: |-
        Newest first. ADMIN: all orders; CASHIER: only orders they created, across all their shifts (BR-ROLE-04, OQ-11); a
        CASHIER passing a `cashier_id` other than their own, or a `shift_id` of another cashier's shift → `FORBIDDEN_ROLE`. Filters: `date_from`/`date_to` on `created_at`
        as Cairo dates (OQ-16; the PWA computes اليوم/أمس/هذا الأسبوع with the week starting Saturday), `cashier_id` (ADMIN),
        `payment_method`, `status`, `shift_id`, `search` = order number with or without `#` (US-15.6). `today_counts` gives the
        footer counts for today regardless of filters (OQ-15).

        BR: BR-ROLE-01, BR-ROLE-04, BR-GEN-02, BR-ORD-10, OQ-11, OQ-15, OQ-16, AC-06.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`.
      x-roles:
      - ADMIN
      - CASHIER
      x-br:
      - BR-ROLE-01
      - BR-ROLE-04
      - BR-GEN-02
      - BR-ORD-10
      - OQ-11
      - OQ-15
      - OQ-16
      - AC-06
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      parameters:
      - $ref: '#/components/parameters/DateFrom'
      - $ref: '#/components/parameters/DateTo'
      - name: cashier_id
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      - name: payment_method
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/PaymentMethod'
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/OrderStatus'
      - name: shift_id
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A page of orders.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
    post:
      operationId: createOrder
      tags:
      - orders
      summary: Confirm a new order — "تأكيد الطلب" (screen 11)
      description: |-
        Creates the order (status OPEN) from the client cart (BR-ORD-01) in one transaction (ADR-0006):
        - caller must have an OPEN shift → else `NO_OPEN_SHIFT` (BR-SHIFT-03); the order belongs to the caller and that shift
          (BR-ORD-03);
        - `lines` empty → `ORDER_EMPTY` (BR-ORD-02); qty integer ≥ 1; note ≤ 140;
        - every item must be ACTIVE in an ACTIVE category at confirm time → else `ITEM_INACTIVE` with `details.item_ids`, no
          order created (BR-ITEM-01, OQ-22); snapshots of name and current price are stored (BR-ITEM-03);
        - lines with the same item and equal note are merged (qty summed) (BR-ORD-05);
        - DINE_IN requires `table_id` of an active table (`TABLE_INACTIVE`) with no OPEN order (`TABLE_OCCUPIED`, enforced by
          the partial unique index under concurrency — BR-TBL-02/03, AC-04); TAKEAWAY must have no table (`VALIDATION_ERROR`)
          (BR-ORD-04); unknown `table_id`/`item_id` → `NOT_FOUND` (BR §9, US-23.8);
        - totals computed by the server (BR-ORD-06, BR-GEN-03); 0 ≤ discount ≤ subtotal else `DISCOUNT_INVALID` (BR-ORD-07);
        - `number` from the global sequence starting at 1001 (BR-ORD-10).
        ADMIN → `FORBIDDEN_ROLE` (BR-ROLE-02). No offline queueing (BR §0).

        Check order (first failure wins): `UNAUTHENTICATED` → `FORBIDDEN_ROLE` → `VALIDATION_ERROR` (body shape, type/table
        pairing) → `NO_OPEN_SHIFT` → `ORDER_EMPTY` → `NOT_FOUND` (table/item ids) → `ITEM_INACTIVE` → `TABLE_INACTIVE` →
        `TABLE_OCCUPIED` → `DISCOUNT_INVALID`.

        BR: BR-ORD-01, BR-ORD-02, BR-ORD-03, BR-ORD-04, BR-ORD-05, BR-ORD-06, BR-ORD-07, BR-ORD-10, BR-SHIFT-03, BR-TBL-02, BR-TBL-03, BR-ITEM-01, BR-ITEM-03, BR-ROLE-02, BR-GEN-03, OQ-22, AC-01, AC-03, AC-04, AC-09, AC-10, AC-11.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NO_OPEN_SHIFT`, `ORDER_EMPTY`, `NOT_FOUND`, `ITEM_INACTIVE`, `TABLE_INACTIVE`, `TABLE_OCCUPIED`, `DISCOUNT_INVALID`, `VALIDATION_ERROR`.
      x-roles:
      - CASHIER
      x-br:
      - BR-ORD-01
      - BR-ORD-02
      - BR-ORD-03
      - BR-ORD-04
      - BR-ORD-05
      - BR-ORD-06
      - BR-ORD-07
      - BR-ORD-10
      - BR-SHIFT-03
      - BR-TBL-02
      - BR-TBL-03
      - BR-ITEM-01
      - BR-ITEM-03
      - BR-ROLE-02
      - BR-GEN-03
      - OQ-22
      - AC-01
      - AC-03
      - AC-04
      - AC-09
      - AC-10
      - AC-11
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NO_OPEN_SHIFT
      - ORDER_EMPTY
      - NOT_FOUND
      - ITEM_INACTIVE
      - TABLE_INACTIVE
      - TABLE_OCCUPIED
      - DISCOUNT_INVALID
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
  /api/orders/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      operationId: getOrder
      tags:
      - orders
      summary: Order details (screens 07, 12, 14)
      description: |-
        Full order with lines (snapshots), totals, payment (if PAID) and cancel info (if CANCELLED). ADMIN: any order,
        read-only. CASHIER: their own orders in any status; another cashier's order only while it is OPEN and DINE_IN (the
        tables flow, read-only — BR-ROLE-06, US-12.9), otherwise `FORBIDDEN_ROLE` (BR-ROLE-04; same rule as another cashier's
        shift, GATE B D1). Unknown id → `NOT_FOUND`.

        BR: BR-ROLE-01, BR-ROLE-04, BR-ROLE-06, BR-ORD-06, BR-ITEM-03, AC-09.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`.
      x-roles:
      - ADMIN
      - CASHIER
      x-br:
      - BR-ROLE-01
      - BR-ROLE-04
      - BR-ROLE-06
      - BR-ORD-06
      - BR-ITEM-03
      - AC-09
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      responses:
        '200':
          description: The order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
    put:
      operationId: updateOrder
      tags:
      - orders
      summary: Edit an OPEN order (screen 11 edit mode)
      description: |-
        Replaces the line list and discount of an OPEN order owned by the caller, in one transaction with the order row locked.
        Checks in order: ADMIN → `FORBIDDEN_ROLE` (BR-ROLE-02); not the owner → `NOT_ORDER_OWNER` (BR-ROLE-06); caller has
        no OPEN shift → `NO_OPEN_SHIFT` (BR-SHIFT-03); not OPEN → `ORDER_NOT_EDITABLE` (BR-ORD-08, AC-05).
        Line semantics: a line with `id` keeps its price/name snapshot and may change qty/note (BR-ITEM-03); lines without `id`
        are new at the current price; a new line merges into an existing one only if item, note and unit price snapshot are
        equal (BR-ORD-05, OQ-21); existing lines not listed are removed. A DISABLED item's existing line may be lowered or
        removed but not raised or re-added → `ITEM_INACTIVE` (OQ-23); changing its note is allowed. Empty list → `ORDER_EMPTY`
        (cancel is a separate operation). Discount rules as on create; never auto-clamped (OQ-20). Type and table cannot change
        (OQ-18): a body containing `type` or `table_id` (or any other unknown field) → `VALIDATION_ERROR`. A line `id` that does
        not belong to this order, or an existing line `id` sent with a different `item_id` → `VALIDATION_ERROR`. An unknown
        `item_id` on a new line → `NOT_FOUND` (BR §9, US-23.8).

        Check order: `UNAUTHENTICATED` → `FORBIDDEN_ROLE` → `NOT_FOUND` (path id) → `VALIDATION_ERROR` (body) →
        `NOT_ORDER_OWNER` → `NO_OPEN_SHIFT` → `ORDER_NOT_EDITABLE` → `ORDER_EMPTY` → `NOT_FOUND` (item ids) → `ITEM_INACTIVE` →
        `DISCOUNT_INVALID`.

        BR: BR-ORD-02, BR-ORD-05, BR-ORD-06, BR-ORD-07, BR-ORD-08, BR-ITEM-01, BR-ITEM-03, BR-ROLE-02, BR-ROLE-06, BR-SHIFT-03, BR-GEN-03, OQ-18, OQ-20, OQ-21, OQ-23, AC-05, AC-09, AC-11.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `NOT_ORDER_OWNER`, `NO_OPEN_SHIFT`, `ORDER_NOT_EDITABLE`, `ORDER_EMPTY`, `ITEM_INACTIVE`, `DISCOUNT_INVALID`, `VALIDATION_ERROR`.
      x-roles:
      - CASHIER
      x-br:
      - BR-ORD-02
      - BR-ORD-05
      - BR-ORD-06
      - BR-ORD-07
      - BR-ORD-08
      - BR-ITEM-01
      - BR-ITEM-03
      - BR-ROLE-02
      - BR-ROLE-06
      - BR-SHIFT-03
      - BR-GEN-03
      - OQ-18
      - OQ-20
      - OQ-21
      - OQ-23
      - AC-05
      - AC-09
      - AC-11
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - NOT_ORDER_OWNER
      - NO_OPEN_SHIFT
      - ORDER_NOT_EDITABLE
      - ORDER_EMPTY
      - ITEM_INACTIVE
      - DISCOUNT_INVALID
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
  /api/orders/{id}/cancel:
    parameters:
    - $ref: '#/components/parameters/Id'
    post:
      operationId: cancelOrder
      tags:
      - orders
      summary: Cancel an OPEN order (screen 12)
      description: |-
        Requires a reason with trimmed length ≥ 3 (else `VALIDATION_ERROR`, BR-ORD-09). Same ownership/shift/state checks as
        updateOrder (row lock; BR-ROLE-06, BR-SHIFT-03, BR-ORD-08). Sets CANCELLED with `cancelled_at` and `cancel_reason`;
        the table becomes AVAILABLE (BR-TBL-05); the order stays in history and is excluded from shift totals (BR-SHIFT-05,
        AC-06). Racing a payment: exactly one wins, the other gets `ORDER_NOT_EDITABLE` (US-24.4).

        Check order: `UNAUTHENTICATED` → `FORBIDDEN_ROLE` → `NOT_FOUND` → `VALIDATION_ERROR` (reason) → `NOT_ORDER_OWNER` →
        `NO_OPEN_SHIFT` → `ORDER_NOT_EDITABLE`.

        BR: BR-ORD-08, BR-ORD-09, BR-TBL-05, BR-SHIFT-03, BR-SHIFT-05, BR-ROLE-02, BR-ROLE-06, AC-05, AC-06, AC-11.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `NOT_ORDER_OWNER`, `NO_OPEN_SHIFT`, `ORDER_NOT_EDITABLE`, `VALIDATION_ERROR`.
      x-roles:
      - CASHIER
      x-br:
      - BR-ORD-08
      - BR-ORD-09
      - BR-TBL-05
      - BR-SHIFT-03
      - BR-SHIFT-05
      - BR-ROLE-02
      - BR-ROLE-06
      - AC-05
      - AC-06
      - AC-11
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - NOT_ORDER_OWNER
      - NO_OPEN_SHIFT
      - ORDER_NOT_EDITABLE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCancel'
      responses:
        '200':
          description: Cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: pos_refresh
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
        maxLength: 100
    DateFrom:
      name: date_from
      in: query
      required: false
      description: Inclusive Cairo calendar date (ADR-0013).
      schema:
        type: string
        format: date
    DateTo:
      name: date_to
      in: query
      required: false
      description: Inclusive Cairo calendar date (ADR-0013).
      schema:
        type: string
        format: date
...
  responses:
    Error401:
      description: INVALID_CREDENTIALS (login) or UNAUTHENTICATED (missing/invalid/expired/revoked session).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: FORBIDDEN_ROLE, NOT_ORDER_OWNER, or USER_DISABLED (login).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: NOT_FOUND.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: State conflict — see the operation's x-error-codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error422:
      description: VALIDATION_ERROR, ORDER_EMPTY, DISCOUNT_INVALID or INSUFFICIENT_CASH.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    CategoryIcon:
      type: string
      description: Fixed built-in icon set (BR-ITEM-07, OQ-33). Keys are rendered by the PWA; no uploads. The 20 keys were
        approved by the owner at GATE B (D4).
      enum:
      - coffee
      - tea
      - juice
      - soft_drink
      - water
      - breakfast
      - sandwich
      - burger
      - pizza
      - grill
      - chicken
      - pasta
      - rice
      - salad
      - soup
      - appetizer
      - dessert
      - cake
      - ice_cream
      - bakery
...
    OrderType:
      type: string
      enum:
      - DINE_IN
      - TAKEAWAY
    OrderStatus:
      type: string
      enum:
      - OPEN
      - PAID
      - CANCELLED
    PaymentMethod:
      type: string
      enum:
      - CASH
      - CARD
      description: BR-PAY-01.
    ErrorCode:
      type: string
      description: BR §9 codes (BR-GEN-06). The HTTP status of each code is fixed by BR §9. `INTERNAL_ERROR` (500) is technical,
        not a business rejection; approved by the owner at GATE B (D3) and added to BR §9 in BR v1.3 (ADR-0010).
      enum:
      - INVALID_CREDENTIALS
      - USER_DISABLED
      - TOO_MANY_ATTEMPTS
      - FORBIDDEN_ROLE
      - SHIFT_ALREADY_OPEN
      - NO_OPEN_SHIFT
      - SHIFT_HAS_OPEN_ORDERS
      - TABLE_OCCUPIED
      - TABLE_INACTIVE
      - ITEM_INACTIVE
      - ORDER_EMPTY
      - ORDER_NOT_EDITABLE
      - NOT_ORDER_OWNER
      - DISCOUNT_INVALID
      - INSUFFICIENT_CASH
      - ITEM_IN_USE
      - CATEGORY_NOT_EMPTY
      - USER_HAS_OPEN_SHIFT
      - LAST_ADMIN
      - VALIDATION_ERROR
      - UNAUTHENTICATED
      - NOT_FOUND
      - DUPLICATE_VALUE
      - IDEMPOTENCY_CONFLICT
      - INTERNAL_ERROR
    Pagination:
      type: object
      required:
      - page
      - page_size
      - total_items
      - total_pages
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 1
    ErrorDetails:
      type: object
      description: Always present, may be empty. Known keys per code (ADR-0010).
      properties:
        fields:
          type: object
          description: 'VALIDATION_ERROR: field path → machine hints (English, never shown).'
          additionalProperties:
            type: array
            items:
              type: string
        open_order_numbers:
          type: array
          items:
            type: integer
          description: SHIFT_HAS_OPEN_ORDERS (BR-SHIFT-04).
        retry_after_seconds:
          type: integer
          nullable: true
          description: TOO_MANY_ATTEMPTS.
        item_ids:
          type: array
          items:
            type: integer
          description: ITEM_INACTIVE (BR-ITEM-01).
        field:
          type: string
          description: 'DUPLICATE_VALUE: the unique field (BR §9).'
      additionalProperties: true
    Error:
      type: object
      required:
      - success
      - code
      - message
      - details
      properties:
        success:
          type: boolean
          enum:
          - false
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Arabic message from BR §9, verbatim.
        details:
          $ref: '#/components/schemas/ErrorDetails'
...
    OrderLine:
      type: object
      required:
      - id
      - item_id
      - item_name_snapshot
      - unit_price_snapshot_minor
      - qty
      - note
      - line_total_minor
      - category_icon
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        item_id:
          type: integer
          format: int64
          minimum: 1
        item_name_snapshot:
          type: string
          description: BR-ITEM-03.
        unit_price_snapshot_minor:
          type: integer
          format: int64
          description: Price at the time the line was added (BR-ITEM-03). Integer minor units (BR-GEN-01).
        qty:
          type: integer
          minimum: 1
        note:
          type: string
          maxLength: 140
        line_total_minor:
          type: integer
          format: int64
          description: unit_price_snapshot × qty (BR-ORD-06). Integer minor units (BR-GEN-01).
        category_icon:
          allOf:
          - $ref: '#/components/schemas/CategoryIcon'
          nullable: true
          description: BR-ITEM-07.
    Payment:
      type: object
      required:
      - id
      - order_id
      - method
      - amount_due_minor
      - amount_received_minor
      - change_minor
      - paid_at
      - cashier_id
      - shift_id
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        order_id:
          type: integer
          format: int64
          minimum: 1
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount_due_minor:
          type: integer
          format: int64
          description: = order total. Integer minor units (BR-GEN-01).
        amount_received_minor:
          type: integer
          format: int64
          description: 'CASH: entered (≥ due); CARD: = due (BR-PAY-02/03). Integer minor units (BR-GEN-01).'
        change_minor:
          type: integer
          format: int64
          description: received − due; 0 for CARD (BR-PAY-02/03). Integer minor units (BR-GEN-01).
        paid_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
        cashier_id:
          type: integer
          format: int64
          minimum: 1
        shift_id:
          type: integer
          format: int64
          minimum: 1
    Order:
      type: object
      required:
      - id
      - number
      - type
      - status
      - table_id
      - table_number
      - cashier_id
      - cashier_name
      - shift_id
      - lines
      - item_count
      - subtotal_minor
      - discount_minor
      - total_minor
      - created_at
      - paid_at
      - cancelled_at
      - cancel_reason
      - payment
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        number:
          type: integer
          minimum: 1001
          description: 'Global sequence, shown as #1048 (BR-ORD-10).'
        type:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        table_id:
          type: integer
          format: int64
          minimum: 1
          nullable: true
        table_number:
          type: integer
          nullable: true
          description: null for TAKEAWAY (سفري).
        cashier_id:
          type: integer
          format: int64
          minimum: 1
        cashier_name:
          type: string
        shift_id:
          type: integer
          format: int64
          minimum: 1
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        item_count:
          type: integer
          description: Σ qty.
        subtotal_minor:
          type: integer
          format: int64
          description: Σ line_total (BR-ORD-06). Integer minor units (BR-GEN-01).
        discount_minor:
          type: integer
          format: int64
          description: Fixed amount, 0 ≤ discount ≤ subtotal (BR-ORD-07). Integer minor units (BR-GEN-01).
        total_minor:
          type: integer
          format: int64
          description: subtotal − discount (BR-ORD-06). Integer minor units (BR-GEN-01).
        created_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
        paid_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
          nullable: true
        cancel_reason:
          type: string
          nullable: true
        payment:
          allOf:
          - $ref: '#/components/schemas/Payment'
          nullable: true
    OrderListItem:
      type: object
      required:
      - id
      - number
      - type
      - status
      - table_number
      - cashier_id
      - cashier_name
      - item_count
      - total_minor
      - payment_method
      - created_at
      - paid_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        number:
          type: integer
        type:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        table_number:
          type: integer
          nullable: true
        cashier_id:
          type: integer
          format: int64
          minimum: 1
        cashier_name:
          type: string
        item_count:
          type: integer
        total_minor:
          type: integer
          format: int64
          description: BR-ORD-06. Integer minor units (BR-GEN-01).
        payment_method:
          allOf:
          - $ref: '#/components/schemas/PaymentMethod'
          nullable: true
          description: null unless PAID.
        created_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
        paid_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
          nullable: true
    OrderDayCounts:
      type: object
      description: Orders created today (Cairo) visible to the caller, by status (screen 15 footer, OQ-15).
      required:
      - total
      - paid
      - open
      - cancelled
      properties:
        total:
          type: integer
        paid:
          type: integer
        open:
          type: integer
        cancelled:
          type: integer
    OrderPage:
      type: object
      required:
      - items
      - pagination
      - today_counts
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderListItem'
        pagination:
          $ref: '#/components/schemas/Pagination'
        today_counts:
          $ref: '#/components/schemas/OrderDayCounts'
    OrderLineInput:
      type: object
      required:
      - item_id
      - qty
      properties:
        item_id:
          type: integer
          format: int64
          minimum: 1
        qty:
          type: integer
          minimum: 1
          maximum: 999
          description: Integer 1–999 (BR-ORD-02; max 999 = GATE B D2).
        note:
          type: string
          maxLength: 140
          default: ''
          description: BR §2 OrderLine (≤ 140).
    OrderCreate:
      type: object
      required:
      - type
      - lines
      properties:
        type:
          $ref: '#/components/schemas/OrderType'
        table_id:
          type: integer
          format: int64
          minimum: 1
          nullable: true
          description: Required iff DINE_IN; must be null/absent for TAKEAWAY.
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineInput'
          description: Empty → ORDER_EMPTY (not VALIDATION_ERROR). Equal item+note lines are merged (BR-ORD-05).
        discount_minor:
          type: integer
          format: int64
          default: 0
          description: Fixed amount; < 0 or > subtotal → DISCOUNT_INVALID (BR-ORD-07). Integer minor units.
    OrderLineUpdateInput:
      type: object
      required:
      - item_id
      - qty
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          nullable: true
          description: Existing line id to keep its snapshot; null/absent = new line.
        item_id:
          type: integer
          format: int64
          minimum: 1
        qty:
          type: integer
          minimum: 1
          maximum: 999
          description: Integer 1–999 (BR-ORD-02; max 999 = GATE B D2).
        note:
          type: string
          maxLength: 140
          default: ''
          description: BR §2 OrderLine (≤ 140).
    OrderUpdate:
      type: object
      required:
      - lines
      - discount_minor
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineUpdateInput'
          description: The complete desired line list. Existing lines not listed are removed.
        discount_minor:
          type: integer
          format: int64
          description: Fixed amount (BR-ORD-07). Never auto-clamped (OQ-20). Integer minor units.
      additionalProperties: false
    OrderCancel:
      type: object
      required:
      - reason
      properties:
        reason:
          type: string
          minLength: 3
          maxLength: 200
          description: Trimmed length 3–200, else VALIDATION_ERROR (BR-ORD-09; max 200 = GATE B D2).
...
    OrderPageEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/OrderPage'
    OrderEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Order'
...